[Improve] refactor clickhouse connector, using Factory to create source/sink
