+++ seatunnel-connectors-v2/connector-clickhouse/src/main/java/org/apache/seatunnel/connectors/seatunnel/clickhouse/source/ClickhouseSource.java
@@ -17,48 +17,20 @@
 
 package org.apache.seatunnel.connectors.seatunnel.clickhouse.source;
 
-import org.apache.seatunnel.shade.com.typesafe.config.Config;
-import org.apache.seatunnel.shade.com.typesafe.config.ConfigFactory;
-
-import org.apache.seatunnel.api.common.PrepareFailException;
-import org.apache.seatunnel.api.common.SeaTunnelAPIErrorCode;
 import org.apache.seatunnel.api.source.Boundedness;
 import org.apache.seatunnel.api.source.SeaTunnelSource;
 import org.apache.seatunnel.api.source.SourceReader;
 import org.apache.seatunnel.api.source.SourceSplitEnumerator;
 import org.apache.seatunnel.api.source.SupportColumnProjection;
 import org.apache.seatunnel.api.source.SupportParallelism;
-import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.api.table.type.SeaTunnelRowType;
-import org.apache.seatunnel.common.config.CheckConfigUtil;
-import org.apache.seatunnel.common.config.CheckResult;
-import org.apache.seatunnel.common.constants.PluginType;
-import org.apache.seatunnel.connectors.seatunnel.clickhouse.exception.ClickhouseConnectorException;
 import org.apache.seatunnel.connectors.seatunnel.clickhouse.state.ClickhouseSourceState;
-import org.apache.seatunnel.connectors.seatunnel.clickhouse.util.ClickhouseUtil;
-import org.apache.seatunnel.connectors.seatunnel.clickhouse.util.TypeConvertUtil;
 
-import com.clickhouse.client.ClickHouseClient;
-import com.clickhouse.client.ClickHouseException;
-import com.clickhouse.client.ClickHouseFormat;
 import com.clickhouse.client.ClickHouseNode;
-import com.clickhouse.client.ClickHouseResponse;
-import com.google.auto.service.AutoService;
-import com.google.common.collect.ImmutableMap;
 
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ThreadLocalRandom;
 
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.DATABASE;
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.HOST;
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.PASSWORD;
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.SERVER_TIME_ZONE;
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.SQL;
-import static org.apache.seatunnel.connectors.seatunnel.clickhouse.config.ClickhouseConfig.USERNAME;
-
-@AutoService(SeaTunnelSource.class)
 public class ClickhouseSource
         implements SeaTunnelSource<SeaTunnelRow, ClickhouseSourceSplit, ClickhouseSourceState>,
                 SupportParallelism,
@@ -68,74 +40,16 @@ public class ClickhouseSource
     private SeaTunnelRowType rowTypeInfo;
     private String sql;
 
+    public ClickhouseSource(
+            List<ClickHouseNode> servers, SeaTunnelRowType rowTypeInfo, String sql) {
+        this.servers = servers;
+        this.rowTypeInfo = rowTypeInfo;
+        this.sql = sql;
+    }
+
     @Override
     public String getPluginName() {
         return "Clickhouse";
-    }
-
-    @Override
-    public void prepare(Config config) throws PrepareFailException {
-        CheckResult result =
-                CheckConfigUtil.checkAllExists(
-                        config,
-                        HOST.key(),
-                        DATABASE.key(),
-                        SQL.key(),
-                        USERNAME.key(),
-                        PASSWORD.key());
-        if (!result.isSuccess()) {
-            throw new ClickhouseConnectorException(
-                    SeaTunnelAPIErrorCode.CONFIG_VALIDATION_FAILED,
-                    String.format(
-                            "PluginName: %s, PluginType: %s, Message: %s",
-                            getPluginName(), PluginType.SOURCE, result.getMsg()));
-        }
-        Map<String, Object> defaultConfig =
-                ImmutableMap.<String, Object>builder()
-                        .put(SERVER_TIME_ZONE.key(), SERVER_TIME_ZONE.defaultValue())
-                        .build();
-
-        config = config.withFallback(ConfigFactory.parseMap(defaultConfig));
-        servers =
-                ClickhouseUtil.createNodes(
-                        config.getString(HOST.key()),
-                        config.getString(DATABASE.key()),
-                        config.getString(SERVER_TIME_ZONE.key()),
-                        config.getString(USERNAME.key()),
-                        config.getString(PASSWORD.key()));
-
-        sql = config.getString(SQL.key());
-        ClickHouseNode currentServer =
-                servers.get(ThreadLocalRandom.current().nextInt(servers.size()));
-        try (ClickHouseClient client = ClickHouseClient.newInstance(currentServer.getProtocol());
-                ClickHouseResponse response =
-                        client.connect(currentServer)
-                                .format(ClickHouseFormat.RowBinaryWithNamesAndTypes)
-                                .query(modifySQLToLimit1(config.getString(SQL.key())))
-                                .executeAndWait()) {
-
-            int columnSize = response.getColumns().size();
-            String[] fieldNames = new String[columnSize];
-            SeaTunnelDataType<?>[] seaTunnelDataTypes = new SeaTunnelDataType[columnSize];
-
-            for (int i = 0; i < columnSize; i++) {
-                fieldNames[i] = response.getColumns().get(i).getColumnName();
-                seaTunnelDataTypes[i] = TypeConvertUtil.convert(response.getColumns().get(i));
-            }
-
-            this.rowTypeInfo = new SeaTunnelRowType(fieldNames, seaTunnelDataTypes);
-
-        } catch (ClickHouseException e) {
-            throw new ClickhouseConnectorException(
-                    SeaTunnelAPIErrorCode.CONFIG_VALIDATION_FAILED,
-                    String.format(
-                            "PluginName: %s, PluginType: %s, Message: %s",
-                            getPluginName(), PluginType.SOURCE, e.getMessage()));
-        }
-    }
-
-    private String modifySQLToLimit1(String sql) {
-        return String.format("SELECT * FROM (%s) s LIMIT 1", sql);
     }
 
     @Override
